Move shaderProgram from HomeScreen to Maze
