+++ src/screens/HomeScreen.cpp
@@ -1,6 +1,5 @@
 #include "HomeScreen.h"
 
-#include "graphics/include/Shader.h"
 #include "graphics/include/config.h"
 #include "graphics/include/utils.h"
 #include "logger/include/logger.h"
@@ -8,10 +7,6 @@
 #include "maze_builder/include/builder.h"
 
 #include "Maze.h"
-
-#include <glm/glm.hpp>
-#include <glm/gtc/matrix_transform.hpp>
-#include <glm/gtc/type_ptr.hpp>
 
 using namespace math;
 
@@ -55,42 +50,11 @@ Maze getMazeFromFile(ushort maze_id, bool join_it, ushort max_lines)
 	return Maze::build(std::move(lines), walls);
 }
 
-std::string getShaderSource(const char* shaderName)
-{
-	return readFile(SHADER_DIRECTORY, shaderName);
-}
-
-void compileShader(Shader& shader)
-{
-	if (!shader.tryToCompile()) {
-		ERR << "\nShader compilation failed: " << shader.infoLog;
-		throw "Shader copmilation fail";
-	}
-}
-
-void createShader(const char* name, GLuint type, const ShaderProgram& program)
-{
-	Shader shader{getShaderSource(name).c_str(), type};
-
-	compileShader(shader);
-	program.attachShader(shader.id);
-}
-
 HomeScreen::HomeScreen(ushort maze_id, bool join_it, ushort max_lines)
     : screen_elements{}
 {
 	screen_elements.emplace_back(
 	    std::make_unique<Maze>(getMazeFromFile(maze_id, join_it, max_lines)));
-	createShader("fragment_shader.glsl", GL_FRAGMENT_SHADER, shaderProgram);
-	createShader("vertex_shader.glsl", GL_VERTEX_SHADER, shaderProgram);
-
-	shaderProgram.link();
-
-	glm::mat4 proj = glm::ortho(0.f, 100.f, 0.f, 100.f, 0.1f, -.1f);
-	const GLfloat* const matrix = glm::value_ptr(proj);
-
-	shaderProgram.use();
-	shaderProgram.setUniformMatrix("model_view_projection", matrix);
 }
 
 void update(const double deltaTime, State& state)
